web socket wrapper deadlock fix
the synchronization was set up because we had crashes from race
conditions in closing the web socket from a thread while using it in
another. this resulted in a deadlock. it doesn't matter if the
messages get sent in this case. so all we need to do is show caution
when using the websocket.